+++ apps/np-atobarai/CHANGELOG.md
@@ -1,5 +1,14 @@
 # saleor-app-payment-np-atobarai
 
+## 1.5.3
+
+### Patch Changes
+
+- ce36410: Upgraded Next.js to v15.5.26
+- Updated dependencies [c12c623]
+  - @saleor/apps-shared@1.17.1
+  - @saleor/apps-trpc@4.0.5
+
 ## 1.5.2
 
 ### Patch Changes