+++ apps/segment/CHANGELOG.md
@@ -1,5 +1,13 @@
 # saleor-app-segment
 
+## 2.4.3
+
+### Patch Changes
+
+- ce36410: Upgraded Next.js to v15.5.26
+- Updated dependencies [c12c623]
+  - @saleor/apps-shared@1.17.1
+
 ## 2.4.2
 
 ### Patch Changes